chore: reorganize module structure
